CAMEL-13449: camel3 - Move bean component out of camel-core
