+++ core/camel-core/src/main/java/org/apache/camel/processor/aggregate/ParameterInfo.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.bean;
+package org.apache.camel.processor.aggregate;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
